requirements.txt
Django==1.7.7
requests==2.6.0
django-autoslug==1.7.2
